Renamed modules to make then consistent.
Remove base sources from test compiles.
