+++ SocketIO-MacTests/SocketAckManagerTest.swift
@@ -7,7 +7,7 @@
 //
 
 import XCTest
-@testable import SocketIOClientSwift
+@testable import SocketIO
 
 class SocketAckManagerTest: XCTestCase {
     var ackManager = SocketAckManager()
